+++ clang/include/clang/AST/Stmt.h
@@ -686,6 +686,9 @@ protected:
 
     unsigned : NumExprBits;
 
+    /// Whether this CXXDefaultArgExpr rewrote its argument and stores a copy.
+    unsigned HasRewrittenInit : 1;
+
     /// The location where the default argument expression was used.
     SourceLocation Loc;
   };
@@ -695,6 +698,10 @@ protected:
     friend class CXXDefaultInitExpr;
 
     unsigned : NumExprBits;
+
+    /// Whether this CXXDefaultInitExprBitfields rewrote its argument and stores
+    /// a copy.
+    unsigned HasRewrittenInit : 1;
 
     /// The location where the default initializer expression was used.
     SourceLocation Loc;